+++ examples/utils/calculateProportionalAmounts.ts
@@ -1,9 +1,17 @@
+/**
+ * Example showing how to remove liquidity from a pool in recovery mode.
+ * (Runs against a local Anvil fork)
+ *
+ * Run with:
+ * pnpm example ./examples/utils/calculateProportionalAmounts.ts
+ */
+
 import { BalancerApi } from '@/data';
 import { calculateProportionalAmounts } from '@/entities/utils/calculateProportionalAmounts';
 import { InputAmount } from '@/types';
 import { ChainId } from '@/utils';
 
-export default async function calculateProportionalAmountsExample() {
+const calculateProportionalAmountsExample = async () => {
     const chainId = ChainId.MAINNET;
     const poolDataProvider = new BalancerApi(
         'https://api-v3.balancer.fi/',
@@ -14,19 +22,23 @@ export default async function calculateProportionalAmountsExample() {
     const pool =
         await poolDataProvider.pools.fetchPoolStateWithBalances(poolId);
     const inputAmount: InputAmount = {
-        rawAmount: BigInt(149277708680793000),
+        rawAmount: 149277708680793000n,
         address: '0x7f39c581f595b53c5cb19bd0b3f8da6c935e2ca0', // wstETH
         decimals: 18,
     };
-    const { amountsIn, bptOut } = calculateProportionalAmounts(
+    const { tokenAmounts, bptAmount } = calculateProportionalAmounts(
         pool,
         inputAmount,
     );
     console.log(
-        `Token Addresses:     ${amountsIn.map(({ address }) => address)}`,
+        `Token Addresses:     ${tokenAmounts.map(({ address }) => address)}`,
     );
     console.log(
-        `Proportional Amounts:${amountsIn.map(({ rawAmount }) => rawAmount)}`,
+        `Proportional Amounts:${tokenAmounts.map(
+            ({ rawAmount }) => rawAmount,
+        )}`,
     );
-    console.log(`BPT Out:             ${bptOut.rawAmount}`);
-}
+    console.log(`BPT Out:             ${bptAmount.rawAmount}`);
+};
+
+export default calculateProportionalAmountsExample;