+++ src/entities/inputValidator/types.ts
@@ -1,7 +1,10 @@
 import { AddLiquidityInput } from '../addLiquidity/types';
 import { CreatePoolInput } from '../createPool/types';
 import { InitPoolInput } from '../initPool/types';
-import { RemoveLiquidityInput } from '../removeLiquidity/types';
+import {
+    RemoveLiquidityInput,
+    RemoveLiquidityRecoveryInput,
+} from '../removeLiquidity/types';
 import { PoolState } from '../types';
 
 export interface InputValidatorBase {
@@ -12,6 +15,11 @@ export interface InputValidatorBase {
 
     validateRemoveLiquidity(
         input: RemoveLiquidityInput,
+        poolState: PoolState,
+    ): void;
+
+    validateRemoveLiquidityRecovery(
+        input: RemoveLiquidityRecoveryInput,
         poolState: PoolState,
     ): void;