+++ src/entities/removeLiquidity/removeLiquidityV2/weighted/removeLiquidityWeighted.ts
@@ -1,22 +1,30 @@
-import { encodeFunctionData } from 'viem';
+import {
+    createPublicClient,
+    encodeFunctionData,
+    formatEther,
+    formatUnits,
+    http,
+} from 'viem';
 import { Token } from '../../../token';
 import { TokenAmount } from '../../../tokenAmount';
 import { WeightedEncoder } from '../../../encoders/weighted';
-import { VAULT, ZERO_ADDRESS } from '../../../../utils/constants';
+import { CHAINS, VAULT, ZERO_ADDRESS } from '../../../../utils/constants';
 import { vaultV2Abi } from '../../../../abi';
 import { parseRemoveLiquidityArgs } from '../../../utils/parseRemoveLiquidityArgs';
 import {
     RemoveLiquidityBase,
     RemoveLiquidityBuildCallOutput,
     RemoveLiquidityInput,
-    RemoveLiquidityKind,
     RemoveLiquidityQueryOutput,
+    RemoveLiquidityRecoveryInput,
 } from '../../types';
-import { PoolState } from '../../../types';
+import { PoolState, PoolStateWithBalances } from '../../../types';
 import { doRemoveLiquidityQuery } from '../../../utils/doRemoveLiquidityQuery';
-import { getSortedTokens } from '../../../utils';
+import { calculateProportionalAmounts, getSortedTokens } from '../../../utils';
 import { getAmountsCall, getAmountsQuery } from '../../helper';
 import { RemoveLiquidityV2BaseBuildCallInput } from '../types';
+import { getPoolTokensV2, getTotalSupply } from '@/utils/tokens';
+import { HumanAmount } from '@/data';
 
 export class RemoveLiquidityWeighted implements RemoveLiquidityBase {
     public async query(
@@ -27,13 +35,10 @@ export class RemoveLiquidityWeighted implements RemoveLiquidityBase {
         const amounts = getAmountsQuery(sortedTokens, input);
 
         const userData = WeightedEncoder.encodeRemoveLiquidityUserData(
-            input.kind === RemoveLiquidityKind.Recovery
-                ? RemoveLiquidityKind.Proportional
-                : input.kind,
+            input.kind,
             amounts,
         );
 
-        // tokensOut will have zero address if removing liquidity to native asset
         const { args, tokensOut } = parseRemoveLiquidityArgs({
             chainId: input.chainId,
             poolId: poolState.id,
@@ -64,6 +69,58 @@ export class RemoveLiquidityWeighted implements RemoveLiquidityBase {
             bptIn,
             amountsOut,
             tokenOutIndex: amounts.tokenOutIndex,
+            vaultVersion: poolState.vaultVersion,
+            chainId: input.chainId,
+        };
+    }
+
+    public async queryRemoveLiquidityRecovery(
+        input: RemoveLiquidityRecoveryInput,
+        poolState: PoolState,
+    ): Promise<RemoveLiquidityQueryOutput> {
+        const client = createPublicClient({
+            transport: http(input.rpcUrl),
+            chain: CHAINS[input.chainId],
+        });
+
+        const sortedTokens = getSortedTokens(poolState.tokens, input.chainId);
+        const [_, tokenBalances] = await getPoolTokensV2(poolState.id, client);
+        const totalShares = await getTotalSupply(poolState.address, client);
+
+        const poolStateWithBalances: PoolStateWithBalances = {
+            ...poolState,
+            tokens: sortedTokens.map((token, i) => ({
+                address: token.address,
+                decimals: token.decimals,
+                index: i,
+                balance: formatUnits(
+                    tokenBalances[i],
+                    token.decimals,
+                ) as HumanAmount,
+            })),
+            totalShares: formatEther(totalShares) as HumanAmount,
+        };
+
+        const { tokenAmounts, bptAmount } = calculateProportionalAmounts(
+            poolStateWithBalances,
+            input.bptIn,
+        );
+        const bptToken = new Token(
+            input.chainId,
+            bptAmount.address,
+            bptAmount.decimals,
+        );
+        const bptIn = TokenAmount.fromRawAmount(bptToken, bptAmount.rawAmount);
+        const amountsOut = tokenAmounts.map((amount, i) =>
+            TokenAmount.fromRawAmount(sortedTokens[i], amount.rawAmount),
+        );
+        return {
+            poolType: poolState.type,
+            removeLiquidityKind: input.kind,
+            poolId: poolState.id,
+            bptIn,
+            amountsOut,
+            tokenOutIndex: undefined,
             vaultVersion: poolState.vaultVersion,
             chainId: input.chainId,
         };